+++ src/modules/esl-base-element/core.ts
@@ -2,6 +2,6 @@ export type {ESLBaseElementShape} from './core/esl-base-element.shape';
 
 export * from './core/esl-base-element';
 
-export * from './decorators/attr';
-export * from './decorators/bool-attr';
-export * from './decorators/json-attr';
+export * from '../esl-utils/decorators/attr';
+export * from '../esl-utils/decorators/bool-attr';
+export * from '../esl-utils/decorators/json-attr';
